+++ clashroyale/official_api/__init__.py
@@ -1,3 +1,2 @@
 from .client import Client
 from .models import *
-from .utils import *