+++ clashroyale/official_api/client.py
@@ -1,5 +1,6 @@
 import asyncio
 import json
+import logging
 from datetime import datetime
 from pathlib import Path
 from urllib.parse import urlencode
@@ -14,6 +15,8 @@ from .models import (BaseAttrDict, PaginatedAttrDict, Cards, Clan, ClanInfo, Cla
 from .utils import API, SqliteDict, clansearch, tournamentsearch, crtag, keys, typecasted
 
 from_timestamp = datetime.fromtimestamp
+
+log = logging.getLogger(__name__)
 
 
 class Client:
@@ -54,6 +57,8 @@ class Client:
     user_agent: Optional[str] = None
         Appends to the default user-agent
     """
+
+    REQUEST_LOG = '{method} {url} has received {text}, has returned {status}'
 
     def __init__(self, token, session=None, is_async=False, **options):
         self.token = token
@@ -109,7 +114,7 @@ class Client:
         return '<OfficialAPI Client async={}>'.format(self.is_async)
 
     def close(self):
-        self.session.close()
+        return self.session.close()
 
     def _raise_for_status(self, resp, text):
         try:
@@ -117,6 +122,7 @@ class Client:
         except json.JSONDecodeError:
             data = text
         code = getattr(resp, 'status', None) or getattr(resp, 'status_code')
+        log.debug(self.REQUEST_LOG.format(method=resp.request_info.method, url=resp.url, text=text, status=code))
         if self.error_debug:
             raise ServerError(resp, data)
         if 300 > code >= 200:  # Request was successful