+++ test/royaleapi/test_blocking.py
@@ -1,3 +1,4 @@
+import logging
 import time
 import unittest
 import os
@@ -6,7 +7,7 @@ import clashroyale
 import requests
 from dotenv import load_dotenv, find_dotenv
 
-load_dotenv(find_dotenv())
+load_dotenv(find_dotenv('../.env'))
 
 TOKEN = os.getenv('royaleapi')
 URL = os.getenv('url', 'https://api.royaleapi.com')
@@ -15,15 +16,12 @@ URL = os.getenv('url', 'https://api.royaleapi.com')
 class TestBlockingClient(unittest.TestCase):
     """Tests all methods in the blocking client that
     uses the `requests` module in `clashroyale`
-
-    Powered by RoyaleAPI
     """
-    def __init__(self, *args, **kwargs):
-        self.clashroyale_client = clashroyale.RoyaleAPI(TOKEN, url=URL, timeout=30)
-        super().__init__(*args, **kwargs)
+    def setUp(self):
+        self.cr = clashroyale.RoyaleAPI(TOKEN, url=URL, timeout=30)
 
-    # PAUSE IN BETWEEN TESTS #
     def tearDown(self):
+        self.cr.close()
         time.sleep(2)
 
     # MISC METHODS #
@@ -31,19 +29,19 @@ class TestBlockingClient(unittest.TestCase):
         """This test will test out:
         - Constants endpoint
         """
-        self.assertTrue(self.clashroyale_client.get_constants().raw_data, dict)
+        self.assertTrue(self.cr.get_constants().raw_data, dict)
 
     def test_get_endpoints(self):
         """This test will test out:
         - Endpoints endpoint
         """
-        self.assertTrue(self.clashroyale_client.get_endpoints(), list)
+        self.assertTrue(self.cr.get_endpoints(), list)
 
     def test_get_version(self):
         """This test will test out:
         - Version endpoint
         """
-        self.assertTrue(self.clashroyale_client.get_version(), str)
+        self.assertTrue(self.cr.get_version(), str)
 
     # PLAYER METHODS #
     def test_get_player(self):
@@ -53,7 +51,7 @@ class TestBlockingClient(unittest.TestCase):
         - Invalid profile fetching
         """
 
-        get_player = self.clashroyale_client.get_player
+        get_player = self.cr.get_player
         tag = '2P0LYQ'
         player = get_player(tag)
         self.assertEqual(player.tag, tag)
@@ -70,7 +68,7 @@ class TestBlockingClient(unittest.TestCase):
         """
 
         tag = '2P0LYQ'
-        battles = self.clashroyale_client.get_player_battles(tag)
+        battles = self.cr.get_player_battles(tag)
         self.assertTrue(isinstance(battles, list))
 
     def test_get_player_chests(self):
@@ -79,32 +77,31 @@ class TestBlockingClient(unittest.TestCase):
         """
 
         tag = '2P0LYQ'
-        chests = self.clashroyale_client.get_player_chests(tag)
+        chests = self.cr.get_player_chests(tag)
         self.assertTrue(isinstance(chests.upcoming, list))
 
-        self.assertTrue(isinstance(chests.super_magical, int) or
-                        chests.super_magical is None)
+        self.assertTrue(isinstance(chests.super_magical, int) or chests.super_magical is None)
 
     def test_get_response(self):
         """This test will test out:
         - BaseAttrDict.response
         """
         tag = '2P0LYQ'
-        chests = self.clashroyale_client.get_player_chests(tag)
+        chests = self.cr.get_player_chests(tag)
         self.assertTrue(isinstance(chests.response, requests.Response))
 
     def test_get_top_players(self):
         """This test will test out:
         - Top players endpoint
         """
-        players = self.clashroyale_client.get_top_players()
+        players = self.cr.get_top_players()
         self.assertTrue(isinstance(players, list))
 
     def test_get_popular_players(self):
         """This test will test out:
         - Popular players endpoint
         """
-        players = self.clashroyale_client.get_popular_players()
+        players = self.cr.get_popular_players()
         self.assertTrue(isinstance(players, list) or isinstance(players, clashroyale.royaleapi.BaseAttrDict))
 
     # ClAN METHODS #
@@ -115,14 +112,14 @@ class TestBlockingClient(unittest.TestCase):
         - Invalid clan fetching
         """
         tag = '29UQQ282'
-        clan = self.clashroyale_client.get_clan(tag)
+        clan = self.cr.get_clan(tag)
         self.assertEqual(clan.tag, tag)
 
         invalid_tag = '293R8FV'
-        self.assertRaises(ValueError, self.clashroyale_client.get_clan, invalid_tag)
+        self.assertRaises(ValueError, self.cr.get_clan, invalid_tag)
 
         invalid_tag = '2P0LYQLYLY20P'
-        self.assertRaises(clashroyale.NotFoundError, self.clashroyale_client.get_clan, invalid_tag)
+        self.assertRaises(clashroyale.NotFoundError, self.cr.get_clan, invalid_tag)
 
     def test_get_clan_battles(self):
         """This test will test out:
@@ -132,13 +129,13 @@ class TestBlockingClient(unittest.TestCase):
         """
 
         tag = '29UQQ282'
-        battles = self.clashroyale_client.get_clan_battles(tag)
+        battles = self.cr.get_clan_battles(tag)
         self.assertTrue(isinstance(battles, list))
         time.sleep(2)
-        battles = self.clashroyale_client.get_clan_battles(tag, type='all')
+        battles = self.cr.get_clan_battles(tag, type='all')
         self.assertTrue(isinstance(battles, list))
         time.sleep(2)
-        battles = self.clashroyale_client.get_clan_battles(tag, type='war')
+        battles = self.cr.get_clan_battles(tag, type='war')
         self.assertTrue(isinstance(battles, list))
 
     def test_get_clan_war(self):
@@ -147,7 +144,7 @@ class TestBlockingClient(unittest.TestCase):
         """
 
         tag = '29UQQ282'
-        clan_war = self.clashroyale_client.get_clan_war(tag)
+        clan_war = self.cr.get_clan_war(tag)
         self.assertTrue(isinstance(clan_war.raw_data, dict))
 
     def test_get_clan_war_log(self):
@@ -156,7 +153,7 @@ class TestBlockingClient(unittest.TestCase):
         """
 
         tag = '29UQQ282'
-        log = self.clashroyale_client.get_clan_war_log(tag)
+        log = self.cr.get_clan_war_log(tag)
         self.assertTrue(isinstance(log, list))
 
     def test_get_clan_history(self):
@@ -165,7 +162,7 @@ class TestBlockingClient(unittest.TestCase):
         - Clan without history enabled history fetching
         """
 
-        get_clan_history = self.clashroyale_client.get_clan_history
+        get_clan_history = self.cr.get_clan_history()
 
         tag = '2U2GGQJ'
         history = get_clan_history(tag)
@@ -180,7 +177,7 @@ class TestBlockingClient(unittest.TestCase):
         - Clan without history enabled history fetching
         """
 
-        get_clan_tracking = self.clashroyale_client.get_clan_tracking
+        get_clan_tracking = self.cr.get_clan_tracking()
 
         tag = '2U2GGQJ'
         history = get_clan_tracking(tag)
@@ -194,21 +191,21 @@ class TestBlockingClient(unittest.TestCase):
         """This test will test out:
         - Tracking clans endpoint
         """
-        clans = self.clashroyale_client.get_tracking_clans()
+        clans = self.cr.get_tracking_clans()
         self.assertTrue(isinstance(clans, list))
 
     def test_get_top_clans(self):
         """This test will test out:
         - Top clans endpoint
         """
-        clans = self.clashroyale_client.get_top_clans()
+        clans = self.cr.get_top_clans()
         self.assertTrue(isinstance(clans, list))
 
     def test_get_popular_clans(self):
         """This test will test out:
         - Popular clans endpoint
         """
-        clans = self.clashroyale_client.get_popular_clans()
+        clans = self.cr.get_popular_clans()
         self.assertTrue(isinstance(clans, list) or isinstance(clans, clashroyale.royaleapi.BaseAttrDict))
 
     # TOURNAMENT METHODS #
@@ -219,34 +216,34 @@ class TestBlockingClient(unittest.TestCase):
         - Invalid tournament fetching
         """
         tag = 'CU2RG8V'
-        player = self.clashroyale_client.get_tournament(tag)
+        player = self.cr.get_tournament(tag)
         self.assertEqual(player.tag, tag)
 
         invalid_tag = '293R8FV'
-        self.assertRaises(ValueError, self.clashroyale_client.get_clan, invalid_tag)
+        self.assertRaises(ValueError, self.cr.get_clan, invalid_tag)
 
         invalid_tag = '2P0LYQLYLY20P'
-        self.assertRaises(clashroyale.NotFoundError, self.clashroyale_client.get_clan, invalid_tag)
+        self.assertRaises(clashroyale.NotFoundError, self.cr.get_clan, invalid_tag)
 
     def test_get_known_tournaments(self):
         """This test will test out:
         - Known tournaments endpoint
         """
-        tournaments = self.clashroyale_client.get_known_tournaments()
+        tournaments = self.cr.get_known_tournaments()
         self.assertTrue(isinstance(tournaments, list))
 
     def test_get_open_tournaments(self):
         """This test will test out:
         - Open tournaments endpoint
         """
-        tournaments = self.clashroyale_client.get_open_tournaments()
+        tournaments = self.cr.get_open_tournaments()
         self.assertTrue(isinstance(tournaments, list))
 
     def test_get_popular_tournaments(self):
         """This test will test out:
         - Popular tournaments endpoint
         """
-        tournaments = self.clashroyale_client.get_popular_tournaments()
+        tournaments = self.cr.get_popular_tournaments()
         self.assertTrue(isinstance(tournaments, list) or isinstance(tournaments, clashroyale.royaleapi.BaseAttrDict))
 
     # DECKS METHODS #
@@ -254,8 +251,15 @@ class TestBlockingClient(unittest.TestCase):
         """This test will test out:
         - Popular decks endpoint
         """
-        decks = self.clashroyale_client.get_popular_decks()
+        decks = self.cr.get_popular_decks()
         self.assertTrue(isinstance(decks, list) or isinstance(decks, clashroyale.royaleapi.BaseAttrDict))
+
+    # OTHERS #
+    def test_logging(self):
+        logger = 'clashroyale.royaleapi.client'
+        with self.assertLogs(logger=logger, level=logging.DEBUG) as cm:
+            self.cr.get_player(self.player_tags[0])
+        self.assertEqual(len(cm.output, 1))
 
 
 if __name__ == '__main__':